Some simple way to smoketest this plugin.
Just a convenience bunch of scripts allowing for a rapid "integration"
smoke-test of the plugin to be sure it is actually producing a valid zip
file in your environment.
This can be run using the following

ruby src/test/ruby/test.rb

and will output a detailed info on the files and content returned by the
plugin.

Change-Id: I48ea6c8c253e026a3852239f044e36fa35ff9758